Compatibility: transparent GUI rendering for pre-3.2 games
Before 3.2.0 in 32-bit games GUIs with transparent background were rendered as if they do not have alpha channel. Since 3.2.0 they are considered having alpha-channel, which broke the "Classic" gui rendering mode.
This commit restores backwards-compatibility with pre-3.2 games.
