+++ Engine/main/game_file.cpp
@@ -614,13 +614,6 @@ int load_game_file() {
         game.options[OPT_GLOBALTALKANIMSPD] = game.options[OPT_GLOBALTALKANIMSPD] != 0 ? 5 : (-5 - 1);
     }
 
-    // 3.20: Fixed GUI AdditiveOpacity mode not working properly if you tried to have a non-alpha sprite on an alpha GUI
-    if (loaded_game_file_version < kGameVersion_320)
-    {
-        // Force new style rendering for gui sprites with alpha channel
-        game.options[OPT_NEWGUIALPHA] = kGuiAlphaRender_AdditiveOpacitySrcCopy;
-    }
-
     if (game.numfonts > MAX_FONTS)
         quit("!This game requires a newer version of AGS. Too many fonts for this version to handle.");
